+++ packages/agents-openai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @openai/agents-openai
 
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies
+  - computer-agents-core@0.4.6
+
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies
+  - computer-agents-core@0.4.5
+
 ## 0.4.4
 
 ### Patch Changes